+++ javascript/node/selenium-webdriver/index.js
@@ -31,7 +31,7 @@ exports.Session = base.require('webdriver.Session');
 exports.WebDriver = base.require('webdriver.WebDriver');
 exports.WebElement = base.require('webdriver.WebElement');
 
-var submodules = {
+var closureModules = {
   Browser: base.require('webdriver.Browser'),
   Capability: base.require('webdriver.Capability'),
   CommandName: base.require('webdriver.CommandName'),
@@ -50,13 +50,27 @@ var submodules = {
   http: require('./http'),
   logging: base.exportPublicApi('webdriver.logging'),
   promise: base.exportPublicApi('webdriver.promise'),
+  proxy: require('./proxy'),
   stacktrace: base.exportPublicApi('webdriver.stacktrace')
 };
 
 
-Object.keys(submodules).forEach(function(key) {
+Object.keys(closureModules).forEach(function(key) {
   exports.__defineGetter__(key, function() {
-    return submodules[key];
+    return closureModules[key];
+  });
+});
+
+
+var subModules = {
+  http: './http',
+  proxy: './proxy'
+};
+
+
+Object.keys(subModules).forEach(function(key) {
+  exports.__defineGetter__(key, function() {
+    return require(subModules[key]);
   });
 });